+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1 ##
+
+1.  Fix: Error if the target component is missing for component icons
+2.  Add: Allow setup a personal config which does not get mixed with team config. Saved under `Library/PersonalHierarchyConfig.asset`
+3.  Change: GameObject Enabled Checker now only displays when any parent gameObjects is disabled
+
 ## 1.0.0 ##
 
 1.  Better background color to not conflict with default one